Moving session-related helpers over to the agent module
