+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/session/SessionIdTimeoutHandler.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android;
+package io.opentelemetry.android.agent.session;
 
 import io.opentelemetry.android.internal.services.applifecycle.ApplicationStateListener;
 import io.opentelemetry.sdk.common.Clock;
