+++ android-agent/src/main/kotlin/io/opentelemetry/android/agent/session/SessionStorage.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.android.session
+package io.opentelemetry.android.agent.session
 
 interface SessionStorage {
     fun get(): Session
